+++ edc-extensions/iatp/tx-iatp-sts-dim/src/test/java/org/eclipse/tractusx/edc/iam/dcp/sts/RemoteTokenServiceClientExtensionTest.java
@@ -1,5 +1,5 @@
-/********************************************************************************
- * Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+/*
+ * Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -15,29 +15,27 @@
  * under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
+ */
 
-package org.eclipse.tractusx.edc.iam.iatp.sts.dim;
+package org.eclipse.tractusx.edc.iam.dcp.sts;
 
+import org.eclipse.edc.iam.identitytrust.sts.remote.RemoteSecureTokenService;
 import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
-import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.spi.system.configuration.Config;
-import org.eclipse.tractusx.edc.iam.dcp.sts.RemoteTokenServiceClientExtension;
 import org.eclipse.tractusx.edc.iam.dcp.sts.dim.DimSecureTokenService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.eclipse.tractusx.edc.iam.dcp.sts.RemoteTokenServiceClientExtension.DIM_URL;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(DependencyInjectionExtension.class)
-public class DimSecureServiceExtensionTest {
+public class RemoteTokenServiceClientExtensionTest {
 
     @Test
     void initialize(ServiceExtensionContext context, RemoteTokenServiceClientExtension extension) {
@@ -48,14 +46,13 @@ public class DimSecureServiceExtensionTest {
     }
 
     @Test
-    void initialize_shouldNotThrow_whenUrlIsMissing(ServiceExtensionContext context, RemoteTokenServiceClientExtension extension) {
+    void initialize_whenUrlIsMissing_fallsBackToRemoteSts(ServiceExtensionContext context, RemoteTokenServiceClientExtension extension) {
         var monitor = context.getMonitor();
         var prefixeMonitor = mock(Monitor.class);
         when(monitor.withPrefix(anyString())).thenReturn(prefixeMonitor);
 
-        assertThatThrownBy(() -> extension.secureTokenService(context))
-                .isInstanceOf(EdcException.class)
-                .hasMessage("No setting found for key edc.iam.sts.dim.url");
+        assertThat(extension.secureTokenService(context))
+                .isInstanceOf(RemoteSecureTokenService.class);
     }
 
 }
